refactor(client-side): change answer reading protocol

+++ web-interface/src/main/webapp/js/Interviewer.js
@@ -4,21 +4,26 @@ class Interviewer {
 
     static updateTableValues(URL, tableId, cookie) {
         let xmlHttp = new XMLHttpRequest();
+        let dateVariableFormat = "last-updating-date=";
         xmlHttp.onload = function () {
             let response = xmlHttp.responseText;
             if (response.trim() !== "") {
-                let responseData = response.split(dateVariableFormat)[1].split("&");
-                Interviewer.lastUpdatingDate = Date.parse(responseData[0]);
-                Interviewer.deleteOldTableValues(tableId);
-                let tableValues = responseData[1];
-                document.getElementById(tableId).append(tableValues);
+                if (response.startsWith(dateVariableFormat)) {
+                    let responseDate = response.split(dateVariableFormat)[1].split("&")[0];
+                    Interviewer.lastUpdatingDate = moment(responseDate, "EEE MMM dd HH:mm:ss zzz yyyy");
+                    Interviewer.deleteOldTableValues(tableId);
+                    let responseTable = response.split("&");
+                    if (responseTable.length === 2) {
+                        let tableValues = responseTable[1].split("response-table=")[1];
+                        document.getElementById(tableId).append(tableValues);
+                    }
+                }
             }
         };
         xmlHttp.open("POST", URL, true);
         xmlHttp.setRequestHeader("Content-Type", "text/plain; charset=UTF-8");
         xmlHttp.setRequestHeader("Cookie", cookie);
-        let dateVariableFormat = "last-updating-date=";
-        xmlHttp.send(dateVariableFormat + Interviewer.lastUpdatingDate.toString());
+        xmlHttp.send(dateVariableFormat + Interviewer.lastUpdatingDate.millisecond);
     }
 
     static deleteOldTableValues(tableId) {
